#1858: address PR comments
